Cleanup: Explicitly use DBIx::Class::Schema::Loader:
In Dancer2-Plugin-DBIC, DBIx::Class::Schema::Loader is mentioned
as a requirement. It's a lazy optional one. It's also hard to spot
because it's being loaded as a string using Module::Load.

Instead, this makes it a bit clearer by using 'require' directly.

It would be better to make this an optional requirement instead of
a default one.