Add support for schema configuration in schema() call:
What troubles me is that this module is not reentrant. You will
always get the same schema object and you can only configure it
once. Configuring it once is great, but having to argue with the
code over at which point to do it is annoying.

(Example: Issue #12 in Dancer2-Plugin-DBIC.)

This now allows the schema() function to also receive parameters
to use instead of the default configuration. It tries todo it in
a non-intrusive way for the code with minimal code changes and
without altering anything outside this function (like the global
configuration).