Retrieve the options before fetching the schema:
The previous logic would try to look up the schema and return it,
and if it didn't exist, it would fetch the options and try to create
one. This makes sense.

However, this also means that if we want to create a schema without
rummaging in the cache, you will have to check whether to even check
the cache.

Instead, we simply retrieve the relevant information for creation
before deciding what to do. Then we could either create a new one
with this configuration or return one.